authentication and athorization using jwt

+++ controllers/api/v1/posts_api.js
@@ -20,14 +20,20 @@ module.exports.index = async function(req, res){
 module.exports.destroy = async function (req, res){
     try {
         let post = await Post.findById(req.params.id);
+        
+        if(post.user == req.user.id){
+            post.remove();
 
-        post.remove();
+            let comments = await Comment.deleteMany({ post: req.params.id });
 
-        let comments = await Comment.deleteMany({ post: req.params.id });
-
-        return res.status(200).json({
-            message: "Post with associated comments deleted"
-        });
+            return res.status(200).json({
+                message: "Post with associated comments deleted"
+            });
+        }else{
+            return res.status(401).json({
+                message: "You cannot delete this post"
+            });
+        }
     } catch (e) {
         return res.status(500).json({
             message: "Internal Server Error"
